refactor : 랭킹 및 백분위 시각화를 위한 아이콘과 색상 결정 함수 분리
